+++ include/readdy/kernel/singlecpu/programs/SCPUCalculateForces.h
@@ -32,7 +32,7 @@
 #ifndef READDY_MAIN_SINGLECPUCALCULATEFORCES_H
 #define READDY_MAIN_SINGLECPUCALCULATEFORCES_H
 
-#include <readdy/model/programs/Actions.h>
+#include <readdy/model/actions/Actions.h>
 #include <readdy/kernel/singlecpu/SCPUKernel.h>
 
 namespace readdy {
