+++ include/readdy/kernel/singlecpu/programs/SCPUUpdateNeighborList.h
@@ -32,7 +32,7 @@
 #ifndef READDY_MAIN_SINGLECPUUPDATENEIGHBORLIST_H
 #define READDY_MAIN_SINGLECPUUPDATENEIGHBORLIST_H
 
-#include <readdy/model/programs/Actions.h>
+#include <readdy/model/actions/Actions.h>
 #include <readdy/kernel/singlecpu/SCPUKernel.h>
 
 namespace readdy {
