+++ kernels/cpu/include/readdy/kernel/cpu/actions/CPUCalculateForces.h
@@ -33,7 +33,7 @@
 #ifndef READDY_CPUKERNEL_CALCULATEFORCES_H
 #define READDY_CPUKERNEL_CALCULATEFORCES_H
 
-#include <readdy/model/programs/Actions.h>
+#include <readdy/model/actions/Actions.h>
 #include <readdy/kernel/cpu/CPUKernel.h>
 
 namespace readdy {
