+++ kernels/cpu_dense/include/readdy/kernel/cpu_dense/actions/reactions/CPUDGillespieParallel.h
@@ -33,7 +33,7 @@
 #define READDY_DENSE_GILLESPIEPARALLEL_H
 
 
-#include <readdy/model/programs/Actions.h>
+#include <readdy/model/actions/Actions.h>
 #include <readdy/kernel/cpu_dense/CPUDKernel.h>
 #include <readdy/kernel/singlecpu/programs/SCPUReactionImpls.h>
 #include "ReactionUtils.h"
